requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
python-dotenv
httpx
